backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
Pillow>=10.4.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
cloudinary==1.41.0
gunicorn==21.2.0
